requirements.txt
absl-py==2.1.0
aiohttp==3.9.5
aiosignal==1.3.1
aniso8601==9.0.1
annotated-types==0.6.0
anyio==4.3.0
appnope==0.1.4
asttokens==2.4.1
astunparse==1.6.3
async-timeout==4.0.3
attrs==23.1.0
bcrypt==4.1.3
beautifulsoup4==4.12.3
bleach==6.1.0
blinker==1.7.0
blis==0.7.11
cachetools==5.3.3
catalogue==2.0.10
certifi==2024.2.2
cffi==1.16.0
chardet==3.0.4
charset-normalizer==3.3.2
click==8.1.7
cloudpathlib==0.16.0
confection==0.1.4
cryptography==42.0.7
cymem==2.0.8
decorator==5.1.1
defusedxml==0.7.1
distro==1.9.0
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.1/en_core_web_sm-3.7.1-py3-none-any.whl
entrypoints==0.4
exceptiongroup==1.2.1
executing==2.0.1
fastjsonschema==2.19.1
filelock==3.14.0
Flask==3.0.0
flask-restplus==0.13.0
flatbuffers==24.3.25
frozenlist==1.4.1
fsspec==2024.3.1
gast==0.5.4
genai==2.1.0
google-ai-generativelanguage==0.6.3
google-api-core==2.19.0
google-api-python-client==2.129.0
google-auth==2.29.0
google-auth-httplib2==0.2.0
google-generativeai==0.5.3
google-pasta==0.2.0
googleapis-common-protos==1.63.0
gpt-2-simple==0.8.1
grpcio==1.63.0
grpcio-status==1.62.2
h11==0.14.0
h5py==3.11.0
httpcore==1.0.5
httplib2==0.22.0
httpx==0.27.0
huggingface-hub==0.23.0
idna==2.8
importlib-metadata==7.0.0
ipython==8.18.1
ipython-genutils==0.2.0
itsdangerous==2.1.2
jedi==0.19.1
Jinja2==3.1.2
joblib==1.4.2
jsonschema==4.20.0
jsonschema-specifications==2023.11.2
jupyter_core==5.7.2
jupyterlab_pygments==0.3.0
keras==3.3.3
langcodes==3.4.0
language_data==1.2.0
libclang==18.1.1
marisa-trie==1.1.1
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib-inline==0.1.7
mdurl==0.1.2
mistune==3.0.2
ml-dtypes==0.3.2
mpmath==1.3.0
multidict==6.0.5
murmurhash==1.0.10
namex==0.0.8
nest-asyncio==1.6.0
networkx==3.2.1
nltk==3.8.1
numpy==1.26.4
openai==0.27.10
opt-einsum==3.3.0
optree==0.11.0
packaging==24.0
pandas==2.2.2
pandocfilters==1.5.1
paramiko==3.4.0
parso==0.8.4
pexpect==4.9.0
pickleshare==0.7.5
pillow==10.3.0
platformdirs==4.2.2
portpicker==1.2.0
preshed==3.0.9
prompt-toolkit==3.0.41
proto-plus==1.23.0
protobuf==4.25.3
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pydantic==2.7.1
pydantic_core==2.18.2
Pygments==2.18.0
PyNaCl==1.5.0
pyparsing==2.4.7
pyrsistent==0.18.1
python-crfsuite==0.9.10
python-dateutil==2.8.2
pytz==2022.4
PyYAML==6.0.1
pyzmq==26.0.3
rake-nltk==1.0.6
redis==5.0.4
referencing==0.32.0
regex==2024.5.10
requests>=2.13.0,<3.0.0
rich==13.7.1
rpds-py==0.13.2
rsa==4.9
safetensors==0.4.3
scikit-learn==1.4.2
scipy==1.13.0
simplegeneric==0.8.1
six==1.12.0
smart-open==6.4.0
sniffio==1.3.1
soupsieve==2.5
spacy==3.7.4
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==1.3.24
srsly==2.4.8
stack-data==0.6.3
sympy==1.12
tabulate==0.9.0
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorflow==2.16.1
tensorflow-io-gcs-filesystem==0.37.0
termcolor==2.4.0
terminado==0.13.3
thinc==8.2.3
threadpoolctl==3.5.0
tiktoken==0.3.3
tinycss2==1.3.0
tokenizers==0.19.1
toposort==1.10
torch==2.3.0
torchaudio==2.3.0
torchvision==0.18.0
tornado==4.5.3
tqdm==4.66.4
traitlets==5.14.3
transformers==4.40.2
typer==0.9.4
typing_extensions==4.11.0
tzdata==2024.1
underthesea==6.8.0
underthesea_core==1.0.4
uritemplate==4.1.1
urllib3==1.24.3
wasabi==1.1.2
wcwidth==0.2.13
weasel==0.3.4
webencodings==0.5.1
Werkzeug==3.0.3
wrapt==1.16.0
yarl==1.9.4
zipp==3.17.0
flask-cors==3.0.10